+++ qiskit/pulse/__init__.py
@@ -6,7 +6,7 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 """Module for Pulses."""
-from .device import DeviceSpecification
+from .channels import DeviceSpecification
 from .commands import (Acquire, FrameChange, PersistentValue, SamplePulse, Snapshot,
                        Kernel, Discriminator, functional_pulse)
 from .configuration import LoConfig
