+++ qiskit/pulse/channels/__init__.py
@@ -5,7 +5,7 @@
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
 
-"""Channel classes for pulse."""
+"""Device-related classes for pulse."""
 
 from .device_specification import DeviceSpecification
 from .pulse_channels import DriveChannel, ControlChannel, MeasureChannel
