+++ qiskit/pulse/channels/channels.py
