+++ qiskit/pulse/channels/pulse_channels.py
