+++ qiskit/pulse/channels/qubit.py
